+++ base/uk.ac.stfc.isis.ibex.dae/src/uk/ac/stfc/isis/ibex/dae/detectordiagnostics/SpectrumInformation.java
@@ -24,7 +24,7 @@ package uk.ac.stfc.isis.ibex.dae.detectordiagnostics;
 import uk.ac.stfc.isis.ibex.model.ModelObject;
 
 /**
- *
+ * The Class SpectrumInformation.
  */
 public class SpectrumInformation extends ModelObject {
     
@@ -33,24 +33,46 @@ public class SpectrumInformation extends ModelObject {
     private Integer integral;
     private Integer maxSpecBinCount;
     
+    /**
+     * Gets the spectrum number.
+     *
+     * @return the spectrum number
+     */
     public Integer getSpectrumNumber() {
         return spectrumNumber;
     }
     
+    /**
+     * Gets the count rate.
+     *
+     * @return the count rate
+     */
     public Double getCountRate() {
         return countRate;
     }
     
+    /**
+     * Gets the integral.
+     *
+     * @return the integral
+     */
     public Integer getIntegral() {
         return integral;
     }
     
+    /**
+     * Gets the max spec bin count.
+     *
+     * @return the max spec bin count
+     */
     public Integer getMaxSpecBinCount() {
         return maxSpecBinCount;
     }
     
     /**
-     * @param value
+     * Sets the spectrum number.
+     *
+     * @param value the new spectrum number
      */
     public void setSpectrumNumber(Integer value) {
         firePropertyChange("spectrumNumber", this.spectrumNumber, this.spectrumNumber = value);
@@ -58,21 +80,27 @@ public class SpectrumInformation extends ModelObject {
     
 
     /**
-     * @param value
+     * Sets the count rate.
+     *
+     * @param value the new count rate
      */
     public void setCountRate(Double value) {
         firePropertyChange("countRate", this.countRate, this.countRate = value);
     }
     
     /**
-     * @param value
+     * Sets the max spec bin count.
+     *
+     * @param value the new max spec bin count
      */
     public void setMaxSpecBinCount(Integer value) {
         firePropertyChange("maxSpecBinCount", this.maxSpecBinCount, this.maxSpecBinCount = value);
     }
     
     /**
-     * @param value
+     * Sets the integral.
+     *
+     * @param value the new integral
      */
     public void setIntegral(Integer value) {
         firePropertyChange("maxSpecBinCount", this.integral, this.integral = value);
